svc & web layers implementation

+++ src/main/java/com/example/jpa/repositories/RoleRepo.java
@@ -3,5 +3,5 @@ package com.example.jpa.repositories;
 import com.example.jpa.entities.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepo extends JpaRepository<Role, Integer> {
+public interface RoleRepo extends JpaRepository<Role, Long> {
 }
